refactor(games-list): extract favorites/reorder into a part file
Phase 3 / A: second part/extension split of `_SystemGamesListState`. Moved the
favorite-toggle + list re-sort methods verbatim into new
`my_games_list/favorites_reorder.dart` as
`extension _FavoritesReorder on _SystemGamesListState`:
- `_toggleFavorite`
- `_reorderGamesListKeepingVisualPosition`
- `_reorderGamesListFollowingGame`

Zero behaviour / public-API change. All state stays on the host State. Two
extension-adaptation idioms (both behaviourally identical):
- the 4 `setState` calls route through the host `rebuild()` bridge (#143);
- `_log` is a `static` field on the host, so from an extension it must be
  qualified as `_SystemGamesListState._log` (unqualified_reference_to_static_
  member_of_extended_type).

Host: 1 insertion (part directive) / 141 deletions. format clean, `flutter
analyze` no issues project-wide, 207/207 tests. 🟢 local-gate only (pure move).
